Feat/develop 0.8.2

+++ component/keyple-calypso/src/main/command/sam/builder/security/SamReadEventCounterCmdBuild.h
@@ -55,7 +55,7 @@ public:
      */
     SamReadEventCounterCmdBuild(SamRevision& revision,
                                 SamEventCounterOperationType operationType,
-                                int index);
+                                uint8_t index);
 
 private:
     /**
